requirements.txt
Flask
camel-ai
python-dotenv
Pillow